refactor: Split from main into cpu package
